extract database connection helpers to a module
