+++ spec/models/carto/api_key_spec.rb
@@ -1,10 +1,10 @@
-# encoding: utf-8
-
 require 'spec_helper_min'
 require 'support/helpers'
+require 'helpers/database_connection_helper'
 
 describe Carto::ApiKey do
   include CartoDB::Factories
+  include DatabaseConnectionHelper
 
   def api_key_table_permissions(api_key, schema, table_name)
     api_key.table_permissions_from_db.find do |tp|
@@ -84,23 +84,6 @@ describe Carto::ApiKey do
       type: 'user',
       data: data
     }
-  end
-
-  def with_connection_from_api_key(api_key)
-    user = api_key.user
-
-    options = ::SequelRails.configuration.environment_for(Rails.env).merge(
-      'database' => user.database_name,
-      'username' => api_key.db_role,
-      'password' => api_key.db_password,
-      'host' => user.database_host
-    )
-    connection = ::Sequel.connect(options)
-    begin
-      yield connection
-    ensure
-      connection.disconnect
-    end
   end
 
   shared_examples_for 'api key' do
